feat: 이미지 업데이트 유무에 따른 분기처리

+++ services/user.js
@@ -54,34 +54,41 @@ const userDetail = async ({ id }) => {
   if (userData.length === 0) {
     throw { status: 404, message: '조회된 사용자가 없습니다' };
   }
-  // return rowSplit(userData, ['project']);
   return rowSplit(userData, ['project']);
 };
 
 // 수정 - (이메일, 비밀번호 제외)
-const userUpdate = async ({ id }, updateData, { destination, uploadedFile, path: _tmpPath }) => {
-  try {
-    if (updateData.nickname) {
-      const checkRows = await userDao.checkNickname(updateData.nickname);
-      if (checkRows.length) {
-        throw { status: 400, message: '중복된 닉네임이 존재합니다' };
-      }
+const userUpdate = async ({ id }, updateData, filedata = null) => {
+  if (updateData.nickname) {
+    const checkRows = await userDao.checkNickname(updateData.nickname);
+    if (checkRows.length) {
+      throw { status: 400, message: '중복된 닉네임이 존재합니다' };
     }
-
-    const previousPath = await userDao.getImage(id);
+  }
+  if (filedata) {
+    // 이미지 수정일 경우
+    const { destination, uploadedFile, path: _tmpPath } = filedata;
+    try {
+      const previousPath = await userDao.getImage(id);
+      const updateRows = await userDao.userUpdate(id, updateData);
+      if (updateRows.affectedRows === 0) {
+        throw { status: 404, message: '조회된 사용자가 없습니다' };
+      }
+      // 기존 이미지 삭제, _tmp 네이밍 변경
+      const oldImagePath = path.join(destination, path.basename(previousPath[0].image || 'nullFileName'));
+      const newPath = path.join(destination, uploadedFile.basename);
+      -fs.rename(_tmpPath, newPath, (err) => {});
+      -fs.unlink(oldImagePath, (err) => {});
+    } catch (err) {
+      fs.unlink(_tmpPath, (err) => {});
+      throw err;
+    }
+  } else {
+    // 이미지 수정이 아닐경우
     const updateRows = await userDao.userUpdate(id, updateData);
     if (updateRows.affectedRows === 0) {
       throw { status: 404, message: '조회된 사용자가 없습니다' };
     }
-
-    // 기존 이미지 삭제, _tmp 네이밍 변경
-    const oldImagePath = path.join(destination, path.basename(previousPath[0].image || 'nullFileName'));
-    const newPath = path.join(destination, uploadedFile.basename);
-    -fs.rename(_tmpPath, newPath, (err) => {});
-    -fs.unlink(oldImagePath, (err) => {});
-  } catch (err) {
-    fs.unlink(_tmpPath, (err) => {});
-    throw err;
   }
 };
